fix: make codebase clippy-clean -- zero warnings across all tool files
Eliminate every cargo clippy warning across the entire codebase:
- is_some_and: replace .map_or(false, |v| ...) across all tool files
- char array patterns: .find(['a','b']) replaces multi-arm closures
- clamp(): replace .max(lo).min(hi) chains
- range contains: (lo..hi).contains(&x) replaces manual comparisons
- needless lifetimes: remove explicit lifetimes the compiler can elide
- needless range loops: iter_mut().enumerate() replaces index loops
- ptr_arg: &Path replaces &PathBuf in function signatures
- wrong_self_convention: rename from_base -> unit_from_base
- type_complexity: add VProp type alias in vcf_tools
- to_string_in_format_args: drop .to_string() where Display suffices
- if_same_then_else: collapse identical branches with ||
- only_used_in_recursion: convert collect_words to free function
- map_identity: remove .map(|(v,n)| (v,n)) identity map
- unused imports/variables: remove PathBuf import, prefix _n

+++ src/tools/audio_file_tools.rs
@@ -61,6 +61,7 @@ fn ru24be(b: &[u8], o: usize) -> u32 {
     }
     ((b[o] as u32) << 16) | ((b[o + 1] as u32) << 8) | (b[o + 2] as u32)
 }
+#[allow(dead_code)]
 fn ru24le(b: &[u8], o: usize) -> u32 {
     if o + 2 >= b.len() {
         return 0;
@@ -645,7 +646,7 @@ fn parse_flac(b: &[u8]) -> Option<FlacInfo> {
                     // bits 0-19 = sample_rate, bits 20-22 = channels-1, bits 23-27 = bps-1, bits 28-27+36 = total_samples
                     let sr = ((d[10] as u32) << 12) | ((d[11] as u32) << 4) | ((d[12] as u32) >> 4);
                     let channels = ((d[12] >> 1) & 7) + 1;
-                    let bps = (((d[12] & 1) as u8) << 4) | (d[13] >> 4) + 1;
+                    let bps = ((d[12] & 1) << 4) | ((d[13] >> 4) + 1);
                     let ts = (((d[13] & 0xF) as u64) << 32)
                         | ((d[14] as u64) << 24)
                         | ((d[15] as u64) << 16)
